# book-network-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi specification - Abdelrahman magdi",
        "description": "OpenApi documentation for Spring Security",
        "termsOfService": "Terms of service",
        "contact": {
            "name": "Abdelrahman magdi",
            "email": "abdelrahmanmagdi527@gmail.com"
        },
        "license": {
            "name": "Licence name",
            "url": "https://some-url.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api/v1",
            "description": "Local ENV"
        },
        {
            "url": "https://aliboucoding.com/course",
            "description": "PROD ENV"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/feedbacks": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "operationId": "saveFeedback",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "operationId": "request",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegistrationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/authentication": {
            "post": {
                "tags": [
                    "auth"
                ],
                "operationId": "authentication",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/feedbacks/book/{book-id}": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "operationId": "findAllFeedbacksByBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseFeedbackResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "FeedbackRequest": {
                "type": "object",
                "properties": {
                    "note": {
                        "type": "number",
                        "format": "double",
                        "maximum": 5,
                        "minimum": 0
                    },
                    "comment": {
                        "type": "string",
                        "minLength": 1
                    },
                    "bookId": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "required": [
                    "bookId",
                    "comment"
                ]
            },
...
            "RegistrationRequest": {
                "type": "object",
                "properties": {
                    "firstname": {
                        "type": "string",
                        "minLength": 1
                    },
                    "lastname": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 2147483647,
                        "minLength": 8
                    }
                },
                "required": [
                    "email",
                    "firstname",
                    "lastname",
                    "password"
                ]
            },
            "AuthenticationRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 2147483647,
                        "minLength": 8
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "AuthenticationResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
            "FeedbackResponse": {
                "type": "object",
                "properties": {
                    "note": {
                        "type": "number",
                        "format": "double"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "ownFeedback": {
                        "type": "boolean"
                    }
                }
            },
            "PageResponseFeedbackResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FeedbackResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT auth description",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}